requirements.txt
h5py==3.9.0
icecream==2.1.3
imageio==2.34.1
matplotlib==3.5.2
natsort==8.4.0
numba==0.57.0
numpy==1.24.1
open3d==0.17.0
opencv_python==4.7.0.72
pandas==2.0.1
Pillow==10.2.0
plyfile==1.0.2
Requests==2.31.0
scikit_learn==1.2.2
scipy==1.10.1
setuptools==69.2.0
tqdm==4.65.0
trimesh==3.23.5
open_clip_torch>=2.22

